Removes all uses of dispatch_async.
